refactor(http): Update tests for refactor
